refactor: split tree shake rn esm tests
